pyproject.toml
[project]
name = "experimentStash"
version = "0.1.0"
description = "Experiment configuration and tool tracking repository"
authors = [
    {name = "César Miguel Valdez Córdova", email = "cmvcordova@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "omegaconf>=2.3.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=13.0.0"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pre-commit>=3.0.0",
    "pytest>=7.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]



[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]
